fixed seeds... hopfully....again...again

+++ migrations/20170608140707_lessonContents.js
@@ -5,7 +5,7 @@ exports.up = function(knex, Promise) {
     table.integer('lesson_id')
       .references('lessons.id')
       .onDelete('CASCADE');
-    table.integer('order')
+    table.float('order')
     table.string('type')
     table.text('content')
   })
